unsorted segment sum
