+++ tfjs-core/src/ops/segment_ops.ts
@@ -19,55 +19,12 @@ import {ENGINE} from '../engine';
 import {Tensor, Tensor1D} from '../tensor';
 import {convertToTensor} from '../tensor_util_env';
 import {TensorLike} from '../types';
-import {assert, isInt, parseAxisParam} from '../util';
+import {parseAxisParam} from '../util';
 
 import {getUndoAxesPermutation} from './axis_util';
-import {expandDims} from './expand_dims';
-import {greaterEqual} from './greater_equal';
-import {logicalAnd} from './logical_and';
-import {maximum} from './maximum';
 import {op} from './operation';
 import {collectGatherOpShapeInfo} from './segment_util';
-import {ones, scalar, zerosLike} from './tensor_ops';
-import {where} from './where';
-
-/**
- * Computes the sum along segments of a `tf.Tensor`.
- *
- * ```js
- * const x = tf.tensor1d([1, 2, 3, 4]);
- * const segmentIds = tf.tensor1d([1, 2, 0, 1], 'int32');
- * const numSegments = 3;
- *
- * x.unsortedSegmentSum(segmentIds, numSegments).print()
- * //or tf.unsortedSegmentSum(x, segmentIds, numSegments)
- * ```
- * @param x The `tf.Tensor` that will be summed along its segments.
- * @param segmentIds A `tf.Tensor1D` whose rank is equal to the rank of `x`'s
- * dimension along the `axis`.  Maps each element of `x` to a segment.
- * @param numSegments The number of distinct `segmentIds`.
- */
-/** @doc {heading: 'Operations', subheading: 'Segment'} */
-function unsortedSegmentSum_<T extends Tensor>(
-    x: T|TensorLike, segmentIds: Tensor1D|TensorLike, numSegments: number): T {
-  const $x = convertToTensor(x, 'x', 'unsortedSegmentSum');
-  const $segmentIds =
-      convertToTensor(segmentIds, 'segmentIds', 'unsortedSegmentSum', 'int32');
-  assert(isInt(numSegments), () => 'numSegments must be of dtype int');
-
-  const gradFunc = (dy: T, saved: Tensor[]) => {
-    const [$segmentIds] = saved;
-    const derX = () => {
-      return gatherDropNegatives(dy, $segmentIds as Tensor1D);
-    };
-    return {$x: derX};
-  };
-  return ENGINE.runKernelFunc((backend, save) => {
-    const res = backend.unsortedSegmentSum($x, $segmentIds, numSegments);
-    save([$segmentIds]);
-    return res;
-  }, {$x}, gradFunc) as T;
-}
+import {unsortedSegmentSum} from './unsorted_segment_sum';
 
 /**
  * Gather slices from tensor `x`'s axis `axis` according to `indices`.
@@ -158,21 +115,4 @@ function arrayConcat(arrays: number[][]): number[] {
   return result;
 }
 
-function gatherDropNegatives<T extends Tensor>(x: T, indices: Tensor1D) {
-  // Helper function for unsorted segment ops. Gathers params for
-  // positive segment ids and gathers 0 for inputs with negative segment id.
-  // Mirrors _GatherDropNegatives from tensorflow/python/ops/math_grad.py
-  const zeroClippedIndices = maximum(indices, zerosLike(indices));
-  const gathered = gather(x, zeroClippedIndices as Tensor1D);
-  let isPositive = greaterEqual(indices, scalar(0, 'int32'));
-  const numIters = gathered.rank - isPositive.rank;
-  for (let i = 0; i < numIters; ++i) {
-    isPositive = expandDims(isPositive, i + 1);
-  }
-  isPositive = logicalAnd(isPositive, ones(gathered.shape, 'bool'));
-  const zeroSlice = zerosLike(gathered);
-  return where(isPositive, gathered, zeroSlice);
-}
-
 export const gather = op({gather_});
-export const unsortedSegmentSum = op({unsortedSegmentSum_});
