+++ tfjs-core/src/tests.ts
@@ -87,6 +87,7 @@ import './ops/expand_dims_test';
 import './ops/eye_test';
 import './ops/fused_test';
 import './ops/gather_nd_test';
+import './ops/gather_test';
 import './ops/gram_schmidt_test';
 import './ops/greater_equal_test';
 import './ops/greater_test';
@@ -139,7 +140,6 @@ import './ops/reverse_3d_test';
 import './ops/reverse_4d_test';
 import './ops/reverse_test';
 import './ops/scatter_nd_test';
-import './ops/segment_ops_test';
 import './ops/selu_test';
 import './ops/sigmoid_cross_entropy_test';
 import './ops/signal_ops_test';
@@ -159,6 +159,7 @@ import './ops/topk_test';
 import './ops/transpose_test';
 import './ops/truncated_normal_test';
 import './ops/unary_ops_test';
+import './ops/unsorted_segment_sum_test';
 import './ops/unstack_test';
 import './ops/where_async_test';
 import './ops/where_test';
